Remove description len limit for GenericTaskType
The description is stored in a TextField (see
https://docs.djangoproject.com/en/2.0/ref/models/fields/#textfield),
which allows for an arbitrary number of characters.

The max length attribute does not affect the model but simply imposes a
restriction on the length of a corresponding FormField widget, which is
not desirable in this case (the whole point of using TextField was to
allow for an arbitrary number of characters in the description).