Add last updated field to file res and seq dataset
The date chosen to be applied retroactively was 1960-01-01. Just
something far enough back that no one will get confused about it.

Probably a better default value (as suggested by Sophia) would be to
sync it with the "created" field, but this only exists for File
Resources and not for Sequence Datasets.