TANTA-97: refactor API filters
Now all of the filters are in the form of somewhat customizable
class-based FilterSets (cf. filter_fields). Most of this is just
enforcing that change and the tidy up that goes along with that. One
other important change is that for *all* filters on relational fields,
the browsable DRF filter widget will now be a text field (else you have
drop-downs with thousands if items, which is both unusable and slow!).